refactor: vendor structure
