+++ client/packages/cli/src/commands/register/gateway.ts
@@ -1,17 +1,18 @@
 import "@t3rn/types"
 import ora from "ora"
-import { Encodings, ApiPromise, WsProvider } from "@t3rn/sdk"
-import fetch from "node-fetch"
+import { ApiPromise } from "@t3rn/sdk"
 import { createType } from "@t3rn/types"
 import {
   T3rnPrimitivesGatewayVendor,
   T3rnPrimitivesExecutionVendor,
   T3rnAbiRecodeCodec,
 } from "@polkadot/types/lookup"
-import { Config, Gateway } from "@/schemas/setup.ts"
+import { Gateway } from "@/schemas/setup.ts"
 import { colorLogMsg, log } from "@/utils/log.ts"
 import { createCircuitContext } from "@/utils/circuit.ts"
 import { getConfig } from "@/utils/config.ts"
+import { registerSubstrateVerificationVendor } from "./vendors/substrate.ts"
+import { registerEthereumVerificationVendor } from "./vendors/eth.ts"
 
 export const spinner = ora()
 
@@ -119,6 +120,8 @@ const getRegistrationData = (
   switch (gatewayData.registrationData.verificationVendor) {
     case "Rococo":
       return registerSubstrateVerificationVendor(circuit, gatewayData)
+    case "Ethereum":
+      return registerEthereumVerificationVendor(circuit, gatewayData)
     default:
       spinner.fail(
         colorLogMsg(
@@ -127,139 +130,5 @@ const getRegistrationData = (
         )
       )
       return
-  }
-}
-
-export const registerSubstrateVerificationVendor = async (
-  circuit: ApiPromise,
-  gatewayData: Required<Gateway>
-) => {
-  if (!gatewayData.registrationData.parachain) {
-    const target = await ApiPromise.create({
-      provider: new WsProvider(gatewayData.rpc),
-    })
-    return registerRelaychain(circuit, target, gatewayData)
-  }
-
-  return registerParachain(circuit, gatewayData)
-}
-
-const registerRelaychain = async (
-  circuit: ApiPromise,
-  target: ApiPromise,
-  gatewayData: Required<Gateway>
-) => {
-  const portalConsensusData = await fetchPortalConsensusData(
-    circuit,
-    target,
-    gatewayData
-  )
-
-  if (!portalConsensusData) {
-    return
-  }
-
-  const { registrationHeader, authorities, authoritySetId } =
-    portalConsensusData
-
-  spinner.info(
-    colorLogMsg(
-      "INFO",
-      `Registering Block #${registrationHeader.number.toNumber()}`
-    )
-  )
-  spinner.start()
-
-  return circuit
-    .createType("RelaychainRegistrationData", [
-      registrationHeader.toHex(),
-      Array.from(authorities),
-      authoritySetId,
-      gatewayData.registrationData.owner,
-    ])
-    .toHex()
-}
-
-const registerParachain = async (
-  circuit: ApiPromise,
-  gatewayData: Required<Gateway>
-) =>
-  circuit
-    .createType("ParachainRegistrationData", [
-      gatewayData.registrationData.parachain.relayChainId,
-      gatewayData.registrationData.parachain.id,
-    ])
-    .toHex()
-
-const fetchPortalConsensusData = async (
-  circuit: ApiPromise,
-  target: ApiPromise,
-  gatewayData: Required<Gateway>
-) => {
-  const registrationHeight = await fetchLatestAuthoritySetUpdateBlock(
-    gatewayData.subscan
-  )
-
-  if (!registrationHeight) {
-    return
-  }
-
-  const registrationHeader = await target.rpc.chain.getHeader(
-    await target.rpc.chain.getBlockHash(registrationHeight)
-  )
-  const finalityProof = await target.rpc.grandpa.proveFinality(
-    registrationHeight
-  )
-  const authorities =
-    Encodings.Substrate.Decoders.extractAuthoritySetFromFinalityProof(
-      finalityProof
-    )
-  const registratationHeaderHash = await target.rpc.chain.getBlockHash(
-    registrationHeight
-  )
-  const targetAt = await target.at(registratationHeaderHash)
-  const authoritySetId = await targetAt.query.grandpa.currentSetId()
-  return {
-    registrationHeader,
-    authorities: circuit.createType("Vec<AccountId>", authorities),
-    authoritySetId: circuit.createType("SetId", authoritySetId),
-  }
-}
-
-export const fetchLatestAuthoritySetUpdateBlock = async (
-  subscanIndexerApiEndpoint: string
-) => {
-  try {
-    const response = await fetch(
-      subscanIndexerApiEndpoint + "/api/scan/events",
-      {
-        method: "POST",
-        headers: {
-          "Content-Type": "text/json",
-        },
-        body: JSON.stringify({
-          row: 20,
-          page: 0,
-          module: "grandpa",
-          call: "newauthorities",
-        }),
-      }
-    )
-    const responseData = (await response.json()) as {
-      data: {
-        events: Array<{
-          block_num: number
-        }>
-      }
-    }
-
-    if (response.status !== 200) {
-      throw new Error("Subscan indexer API error")
-    }
-
-    return responseData.data.events.map((entry) => entry.block_num)[0]
-  } catch (error) {
-    spinner.fail(colorLogMsg("ERROR", error))
-    spinner.start()
   }
 }
