+++ client/packages/cli/src/commands/register/vendors/eth.ts
@@ -1,0 +1,9 @@
+import { ApiPromise } from "@t3rn/sdk"
+import { Gateway } from "@/schemas/setup.ts"
+
+export const registerEthereumVerificationVendor = async (
+  circuit: ApiPromise,
+  gatewayData: Required<Gateway>
+) => {
+  //
+}
